pyproject.toml
[project]
name = "operant"
version = "0.1.0"
description = "Operant — Generic multi-tenant SaaS Backend API (FastAPI + PostgreSQL + SQLAlchemy + Alembic)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Operant Contributors" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "email-validator>=2.1.1",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.30",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.13.2",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "operant"}

[tool.setuptools.packages.find]
where = ["operant"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"operant/app/api/**/*.py" = ["B008"]
"operant/app/api/deps.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["operant/app/tests"]
addopts = "-q"


